autogen/requirements.txt
pyautogen==0.2.17
openai>=1.3.0
termcolor>=2.3.0
docker>=6.1.3
python-dotenv>=1.0.0
matplotlib>=3.7.0 